requirements.txt
opencv-python==4.5.5.64
onnxruntime-gpu==1.11.1
Pillow==9.1.1
ass==0.5.2
pandas==1.4.2
networkx==2.8
scipy==1.8.0
matplotlib==3.5.2
py2neo==2021.2.3
scikit-learn==1.0.2
plotly==5.7.0
dash-cytoscape==0.3.0
dash==2.3.1
dash-html-components==2.0.0
scikit-image==0.19.2
pycocotools==2.0.4
neo4j==4.4.4
py2neo==2021.2.3
flask-cors==3.0.10
imreg_dft==2.0.0
torch==1.12.0
torchvision==0.13.0
img2vec-pytorch==1.0.1
dataclasses-json==0.5.7